requirements.txt
click==7.1.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
Werkzeug==1.0.1
Flask-Login==0.5.0
oauthlib==3.1.0
python-dotenv==0.13.0
pyopenssl==19.1.0
requests==2.23.0